Added test for get_graph. Reorganized graph reasoning tests.
