+++ deep_causality/tests/types/reasoning_types/causality_graph_reasoning_tests.rs
@@ -2,7 +2,219 @@
 // Copyright (c) "2023" . Marvin Hansen <marvin.hansen@gmail.com> All rights reserved.
 
 use deep_causality::prelude::*;
-use deep_causality::utils::bench_utils_graph;
+use deep_causality::utils::{bench_utils_graph, test_utils};
+
+#[test]
+fn test_reason_all_causes() {
+    let mut g = CausaloidGraph::new();
+
+    // Add root causaloid
+    let root_causaloid = test_utils::get_test_causaloid();
+    let root_index = g.add_root_causaloid(root_causaloid);
+    let contains_root = g.contains_causaloid(root_index);
+    assert!(contains_root);
+
+    // Add causaloid A
+    let causaloid = test_utils::get_test_causaloid();
+    let idx_a = g.add_causaloid(causaloid);
+    let contains_a = g.contains_causaloid(idx_a);
+    assert!(contains_a);
+
+    // Link causaloid A to root causaloid
+    let res = g.add_edge(root_index, idx_a);
+    assert!(res.is_ok());
+
+    let contains_edge = g.contains_edge(root_index, idx_a);
+    assert!(contains_edge);
+
+    // Add causaloid B
+    let causaloid = test_utils::get_test_causaloid();
+    let idx_b = g.add_causaloid(causaloid);
+    let contains_b = g.contains_causaloid(idx_b);
+    assert!(contains_b);
+
+    // Link causaloid B to root causaloid
+    let res = g.add_edge(root_index, idx_b);
+    assert!(res.is_ok());
+
+    let contains_edge = g.contains_edge(root_index, idx_b);
+    assert!(contains_edge);
+
+    // Add causaloid C
+    let causaloid = test_utils::get_test_causaloid();
+    let idx_c = g.add_causaloid(causaloid);
+    let contains_c = g.contains_causaloid(idx_c);
+    assert!(contains_c);
+
+    // Link causaloid C to A
+    let res = g.add_edge(idx_a, idx_c);
+    assert!(res.is_ok());
+
+    let contains_edge = g.contains_edge(idx_a, idx_c);
+    assert!(contains_edge);
+
+    let percent_active = g.percent_active();
+    assert_eq!(percent_active, 0.0);
+
+    let number_active = g.number_active();
+    assert_eq!(number_active, 0.0);
+
+    let all_true = g.all_active();
+    assert!(!all_true);
+
+    let data = [0.99, 0.98, 0.97];
+    let res = g.reason_all_causes(&data, None).unwrap();
+    assert!(res);
+
+    let all_true = g.all_active();
+    assert!(all_true);
+
+    let percent_active = g.percent_active();
+    assert_eq!(percent_active, 100.0);
+
+    let number_active = g.number_active();
+    assert_eq!(number_active, 4.0);
+}
+
+#[test]
+fn test_reason_subgraph_from_cause() {
+    let mut g = CausaloidGraph::new();
+
+    // Add root causaloid
+    let root_causaloid = test_utils::get_test_causaloid();
+    let root_index = g.add_root_causaloid(root_causaloid);
+    let contains_root = g.contains_causaloid(root_index);
+    assert!(contains_root);
+
+    // Add causaloid A
+    let causaloid = test_utils::get_test_causaloid();
+    let idx_a = g.add_causaloid(causaloid);
+    let contains_a = g.contains_causaloid(idx_a);
+    assert!(contains_a);
+
+    // Link causaloid A to root causaloid
+    let res = g.add_edge(root_index, idx_a);
+    assert!(res.is_ok());
+
+    let contains_edge = g.contains_edge(root_index, idx_a);
+    assert!(contains_edge);
+
+    // Add causaloid B
+    let causaloid = test_utils::get_test_causaloid();
+    let idx_b = g.add_causaloid(causaloid);
+    let contains_b = g.contains_causaloid(idx_b);
+    assert!(contains_b);
+
+    // Link causaloid B to root causaloid
+    let res = g.add_edge(root_index, idx_b);
+    assert!(res.is_ok());
+
+    let contains_edge = g.contains_edge(root_index, idx_b);
+    assert!(contains_edge);
+
+    // Add causaloid C
+    let causaloid = test_utils::get_test_causaloid();
+    let idx_c = g.add_causaloid(causaloid);
+    let contains_c = g.contains_causaloid(idx_c);
+    assert!(contains_c);
+
+    // Link causaloid C to A
+    let res = g.add_edge(idx_a, idx_c);
+    assert!(res.is_ok());
+
+    let contains_edge = g.contains_edge(idx_a, idx_c);
+    assert!(contains_edge);
+
+    let percent_active = g.percent_active();
+    assert_eq!(percent_active, 0.0);
+
+    let all_true = g.all_active();
+    assert!(!all_true);
+
+    let data = [0.99, 0.98];
+    let res = g.reason_subgraph_from_cause(idx_a, &data, None).unwrap();
+    assert!(res);
+
+    let percent_active = g.percent_active();
+    assert_eq!(percent_active, 50.0);
+}
+
+#[test]
+fn test_reason_shortest_path_between_causes() {
+    // Reasons over a linear graph:
+    // root(0) -> A(1) -> B(2) -> C(3) ... XYZ(100)
+    // We assume a linear chain of causality.
+    let (g, data) = bench_utils_graph::get_small_linear_graph_and_data();
+
+    // Verify that the graph is fully inactive.
+    let percent_active = g.percent_active();
+    assert_eq!(percent_active, 0.0);
+
+    let number_active = g.number_active();
+    assert_eq!(number_active, 0.0);
+
+    let all_true = g.all_active();
+    assert!(!all_true);
+
+    let start_index = 10;
+    let stop_index = 19;
+    let res = g.reason_shortest_path_between_causes(
+        start_index,
+        stop_index,
+        &data,
+        None,
+    ).unwrap();
+    assert!(res);
+
+    let all_true = g.all_active();
+    assert!(!all_true);
+
+    let number_true = g.number_active();
+    assert_eq!(number_true, 10.0);
+
+    let start_index = 30;
+    let stop_index = 49;
+    let res = g.reason_shortest_path_between_causes(
+        start_index,
+        stop_index,
+        &data,
+        None,
+    ).unwrap();
+    assert!(res);
+
+    let number_true = g.number_active();
+    assert_eq!(number_true, 30.0);
+
+    let all_true = g.all_active();
+    assert!(!all_true);
+}
+
+#[test]
+fn test_reason_single_cause() {
+    let mut g = CausaloidGraph::new();
+    let causaloid = test_utils::get_test_causaloid();
+
+    let index = g.add_causaloid(causaloid);
+    let contains = g.contains_causaloid(index);
+    assert!(contains);
+
+    let percent_active = g.percent_active();
+    assert_eq!(percent_active, 0.0);
+
+    let all_true = g.all_active();
+    assert!(!all_true);
+
+    let obs = 0.99;
+    let res = g.reason_single_cause(index, &[obs]).unwrap();
+
+    assert!(res);
+
+    let percent_active = g.percent_active();
+    assert_eq!(percent_active, 100.0);
+
+    let all_true = g.all_active();
+    assert!(all_true);
+}
 
 #[test]
 fn test_linear_graph() {
